requirements.txt
# ══════════════════════════════════════════════════════
# Polymarket Intelligence System — Python dependencies
# ══════════════════════════════════════════════════════

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Async HTTP
aiohttp>=3.9.0

# Database
duckdb>=0.9.0

# Configuration
python-dotenv>=1.0.0

# Data processing
pyyaml>=6.0

# AI APIs (optional — only needed for Strategy 4)
anthropic>=0.40.0
openai>=1.50.0

# Scheduling
apscheduler>=3.10.0

# Type checking / dev
mypy>=1.7.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
